drv: fix alignment issues on contexts

+++ kernel/src/drv/drv.h
@@ -12,7 +12,7 @@
 
 typedef struct
 {
-    uint32_t pid;
+    uint64_t pid;
     uint8_t keys[64];        // key buffers
     int mouseX, mouseY;      // mouse relative coordonates
     uint8_t mouseButtons[5]; // mouse button states
@@ -20,7 +20,7 @@ typedef struct
 
 typedef struct
 {
-    uint32_t pid; // driver pid
+    uint64_t pid; // driver pid
     void *base;   // base address
     uint32_t currentXres, currentYres;
     uint32_t requestedXres, requestedYres;
