+++ libs/libc/inc/mos/drv.h
@@ -37,7 +37,7 @@ typedef struct
 
 typedef struct
 {
-    uint32_t pid;
+    uint64_t pid;
     uint8_t keys[64];        // key buffers
     int mouseX, mouseY;      // mouse relative coordonates
     uint8_t mouseButtons[5]; // mouse button states
@@ -49,7 +49,7 @@ typedef struct
 
 typedef struct
 {
-    uint32_t pid;
+    uint64_t pid;
     void *base; // base address
     uint32_t currentXres, currentYres;
     uint32_t requestedXres, requestedYres;
